Telegram Bot API June 25, 2021 updates (v5.3)

+++ telegramium-core/src/main/scala/telegramium/bots/client/GetChatMemberReq.scala
@@ -8,4 +8,4 @@ import telegramium.bots.ChatId
   * @param userId
   *   Unique identifier of the target user
   */
-final case class GetChatMemberReq(chatId: ChatId, userId: Int)
+final case class GetChatMemberReq(chatId: ChatId, userId: Long)
